+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 # Unreleased
 
 ### Added
+- Process::info to get debug information on a specific local process.
+- ConsoleServer, enabled using the `console` feature which enables a remote debug api for gathering debug information in realtime.
 
 ### Changed